+++ docs/src/components/ComponentDoc/ComponentDoc.tsx
@@ -3,7 +3,8 @@ import PropTypes from 'prop-types'
 import * as React from 'react'
 import DocumentTitle from 'react-document-title'
 import { withRouter } from 'react-router'
-import { Grid, Header, Icon } from 'semantic-ui-react'
+import { Grid, Icon } from 'semantic-ui-react'
+import { Header } from '@stardust-ui/react'
 
 import componentInfoShape from 'docs/src/utils/componentInfoShape'
 import { scrollToAnchor, examplePathToHash, getFormattedHash } from 'docs/src/utils'
@@ -12,7 +13,7 @@ import ComponentDocSee from './ComponentDocSee'
 import ComponentExamples from './ComponentExamples'
 import ComponentProps from './ComponentProps'
 import ComponentSidebar from './ComponentSidebar'
-import ComponentDocTag from './ComponentDocTag'
+import ComponentAccessibility from './ComponentDocAccessibility'
 
 const topRowStyle = { margin: '1em' }
 const exampleEndStyle: React.CSSProperties = {
@@ -79,18 +80,9 @@ class ComponentDoc extends React.Component<any, any> {
         <Grid>
           <Grid.Row style={topRowStyle}>
             <Grid.Column>
-              <Header
-                as="h1"
-                content={info.displayName}
-                subheader={_.join(info.docblock.description, ' ')}
-              />
-              <ComponentDocTag
-                title="Accessibility"
-                tag="accessibility"
-                errorMessage="No accessibility description found for this component. Please explain the
-                  accessibility concerns of the component by adding the @accessibility tag to the component's docblock."
-                info={info}
-              />
+              <Header as="h1" content={info.displayName} />
+              <p>{_.join(info.docblock.description, ' ')}</p>
+              <ComponentAccessibility info={info} />
               <ComponentDocSee displayName={info.displayName} />
               <ComponentDocLinks
                 displayName={info.displayName}